move Go source to tui/, template/ stays at root via symlink
- All .go files moved from root into tui/
- tui/template is a relative symlink (../template) for navigation
- Build swaps symlink for real copy (cp -rL), builds, restores symlink
  because go:embed cannot follow symlinks
- Removed old superseded tui/ module (tui/go.mod, tui/go.sum, old .go files)
- Makefile and CI updated: build-tui uses ./tui, no more workaround cp
- scripts/maple unchanged — TEMPLATE_DIR still points to root template/
