--- embed.go
@@ -1,6 +1,0 @@
-package main
-
-import "embed"
-
-//go:embed all:template
-var embeddedTemplate embed.FS
