+++ tui/agent_loaders.go
