+++ tui/template
@@ -1,0 +1,1 @@
+../template
