+++ tui/test_discovery.go
